nd_acpi: update calling convention to allow underrun
The BIOS appears to cut its output short on error.  When that happens,
instead of returning -ENXIO, return the positive number of bytes that
remain unwritten in the output buffer.  This assumes that the BIOS has
successfully filled the "status" output field.  The BIOS has filled the
status field when returning at least a 4-byte acpi-buffer object from
acpi_evaluate_dsm().

Also, in debug mode, dump the input buffer, to make it easier to spot
malformed commands from userspace.

Tested-by: Ross Zwisler <ross.zwisler@linux.intel.com>