IB/hfi: Correct issues with processing/setting LRH.SC
STL redefines the LRH.VL field to LRH.SC (lower 4 bits)
and as such it does not hold the VL for the packet any
more.

User level SDMA processing code mistakenly used the
LRH.SC field as the VL for the packet. This results in
errors due to the mismatch between SC and VL number.

Furthermore, (found through code inspection) the Verbs
code was (in some cases) incorrectly setting the LRH.SL
value based on the SC rather than the SL.

This change fixes both of these issue. It also adds an
additional SC/VL verification in the user level SDMA
path to ensure user-passed-in values are sane.

INTERNAL: PR126706 - PSM fails with Unexpected error in
INTERNAL: writev()

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Iae5a381edc760eadc258d2dbca4540cd931c3e17