nd: fix isetcookie calculation
What's wrong with the current algorithm?

The cookie needs to be robust in the case of DIMMs being rearranged in
the system.  Currently we include 'nfit_handle' in the calculation which
will cause our isetcookie to change even though an interleave-set has
maintained integrity after DIMM-slot changes.  The BIOS is the keeper of
"interleave-set integrity", so, we copy its implementation for making
that assessment.  The BIOS does this by writing DIMM serial-numbers in
interleave-set order to the PCD.

What if the DIMM order stays the same, but 'size' or 'DPA-offset' is
changed?

In that case the isetcookie will remain the same, but the 'region_dpa'
and 'region_len' will no longer match what is stored in the label and
the OS can flag the namespace to the administrator.

What about mirrored PMEM?

Tl;dr we may need a separate algorithm for the mirrored case.

In the case of mirrored PMEM the current NFIT definition is
insufficient.  Sammy admits that he needs to spend some time thinking
about how the NFIT can report a degraded-mirror.  My current
recommendation is that an isetcookie is recorded for each interleave-set
in the mirror.  However, we have a problem that the DIMMs that reference
a mirrored SPA may themselves be in separate interleave sets.  In other
words a SPA references 1 set of DIMMs, but that DIMM-set may comprise
multiple interleave sets.  The answer to this problem likely lies in
decoding IDT, but hopefully we can propose NFIT DSM extensions to make
this easier.  Certainly we already need NFIT DSM extensions for mirror
functionality like "fail this mirror leg", "what's the status of
'resilvering'", and as mentioned before "is this mirror degraded?".

Tested-by: Ross Zwisler <ross.zwisler@linux.intel.com>