IB/hfi: fix display of port counters for down HFI port
This patch addresses the pma loopback problem by treating 0xffff as the
loopback LID.

Here is quote for the rationale of the solution summarized in PR 125444:

"The proposal is to allow a GSI application to use a DLID of 0xffff and SLID
of 0xffff in packet it supplied to umad.  This should pass through umad and
ib_mad and ib_core without issue. Within driver, upon receipt of a GSI
packet with DLID of 0xffff, the driver shall loop this packet back to the
MAD stack which will then invoke process_mad as a local request and respond
with a DLID of 0xFFFF which will also then be looped back by the driver.
Note that STL devices will not permit 0xffff DLID on the wire for a non-VL15
packet, so this situation should not occur from the wire, and if such a
packet is received on the wire, the packet handler in WFR should discard it
before delivery to ib_core/ib_mad.  Hence the 0xffff DLID on GSI would be
restricted to local mgmt queries. The main use case is a local PMA query
to access port counters as part of debugging an unstable port. Note on
power up, the default BaseLID for an adapter should be 0.  STL specifically
defines this behavior, while IB was unclear.  However we have not chosen
to use 0 SLID/DLID for this use case because ib_core/ib_mad/umad would
discard such a LID as invalid for IB."

INTERNAL: PR 126401 -  Need a way to display port counters for (up/down) HFI
INTERNAL: port counter
INTERNAL: PR 125444 - Command can't be control-C'd, stops other opasmaquery from
INTERNAL: working, when SM not running
INTERNAL: PR 126402 - port counters reported /sys/class has not been updated for
INTERNAL: STL

Change-Id: I4658d13e19c7abc8f4aeb0a72b16aa4de039e4f0
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Kaike Wan <kaike.wan@intel.com>