IB/hfi: Set default rx context for a0 to 0 for correct RSM match behaviour
The default rx context for RSM is set to invalid (0xff) but
this causes incorrect receive behaviour for A0. So for the
A0 the default rx context is set to zero.

INTERNAL: PR 126639 -- Erratum 291586 - Incorrect receive behaviour
INTERNAL: for RSM match on type 2 (Verbs) or type 4 (bypass) packet
INTERNAL: when the selected receive context is invalid or disabled

Change-Id: I0300c694d01bff28186655ad231f9a9c313b3532
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Sadanand Warrier <sadanand.warrier@intel.com>