IB/hfi: Increase the number of shared contexts
The number of supported shared contexts is increasing
from 4 to 8. This requires that a definition be changed
to reflect the new count.

At the same time, move the new definition to the
exported header file so PSM can start using it and
maintain a matching count.

INTERNAL: PR126831 - Not able to use PSM_SHAREDCONTEXTS_MAX
INTERNAL: variable

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Id3714270ef26b891f257b5b4cdd510dd55ab5089