nd_blk: handle signed int overflow in acquire_bw
atomic_t is a signed value.  This line:

*bw_index = atomic_inc_return(&(dimm->last_bw)) % dimm->num_bw;

will result in a bogus lane number once 2^31 is exceeded, which will
then result in a kernel oops when that number is used as an array index.

This patch makes num_bw an unsigned int.  This causes the int returned
by atomic_int_return() to get promoted to an unsigned int before the mod
operation, preventing the overflow.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Reported-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>