IB/hfi: Add QSFP initialization and interrupt handling support
This patch adds QSFP initialization and interrupt and
error handling. We allow interrupts on cable insert/remove
and on completion of QSFP module reset. On completion of
QSFP module reset, we read the contents of the QSFP
EEPROM and cache it for further use. This patch also provides
a function to acquire QSFP cable information for the CableInfo
SMA query as specified in the IBA 1.3/STL V1G1 specifications.

INTERNAL: PR 125115 -  Use common file format for speciifying
INTERNAL: initial port parameters for LNI as PRR

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>
Change-Id: I392e77d5836d176aec5f7a0ef8f6a816fcafd275