IB/hfi: handle ECOMM failure
The qib driver verbs SDMA send returns 0 when an out of state submission
is attempted and causes a completion to flushed back to the caller as
if the packet was actually sent.

The hfi driver returns -ECOMM and after having scheduled a completion return
for the submited tx requests.

This causes a QP to remain in a busy state.  This inconsistency can lead
to missing send completions causing unload issues.

This patch handles the new -ECOMM returning a 0 to the rest of the
verbs handling as if the packet had been sent.

INTERNAL: PR 125336 - Process hung when unloading the driver after port down
INTERNAL: while SDMA in use

Change-Id: I30f0f51c84b5d9ef90e78cbbae159b9aa18283f2
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>