IB/hfi: drain per-VL FIFOs before modifying certain per-VL resources
When adjusting per-VL resources, it may be necessary to make sure
that the resources are not in use - otherwise packets might get stuck
in a per-VL FIFO with no chance to ever egress. This patch introduces
a mechanism to drain data VLs, and invokes that mechanism when certain
modifications to per-VL resources are made.

There are several per-VL resources which can be modified (e.g., by the
Fabric Manager):

1) buffer control table
2) VL arbitration tables
3) SC/VLt, and SC/VLnt mappings
4) (per-VL) MTU

There are some special considerations for each of these:

1) buffer control table

This should _not_ require quiescing/emptying VL FIFOs, since
there is an existing mechanism (described in section 4.4.7 of
the WFR HAS) to accommodate adjusting the buffer control
table while the link, and all VLs are active.

2) VL arbitration tables

This should only be a problem if the weight for any particular
VL is being set to 0, in which case that VL would need to be
emptied to avoid packets getting "stuck" with no chance to ever
egress.

3) SC/VLt and SC/VLnt mappings

We just can't allow those to change when the link is active.
Changing these mappings must be coordinated between neighbors -
merely draining the VL FIFOs is not sufficient.

4) per-VL MTU

If the MTU for a VL is reduced, it must first be quiesced/drained.


This patch takes the conservative approach, and disables/drains
the per-VL FIFOs whenever VL arbitration tables (2), or per-VL
MTU (4) is changed. Also, note that there's no means to disable,
and drain VL15.

If there is an attempt to send via a disabled VL, the behavior
depends on whether PIO, or SDMA is used. For PIO, we'll get a
disallowed packet error, and the context will be restarted. For
SDMA, we'll get a disallowed packet error, and the packet will
be dropped.

INTERNAL: PR 125823 - mechanism to drain in-flight data for a VL

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Ie87cdf02bbad22b2aba3d2315405ff84958145bd