IB/hfi: don't even try to drain VLs on pre-B0 h/w
commit dd0b52e introduced a problem on pre-B0 h/w, in that the
driver refused to modify the mtu, and VL arbitration tables if
the VL FIFOs could not be drained (i.e., if pre B0 h/w was in
use). This patch preserves the old (pre-dd0b52e) behavior on
pre-B0 h/w, but invokes the drain mechanism otherwise. This
introduces an incompatibility of sorts, since any code which
relies on the drain mechanism (such as a future FM) may not
work properly on pre-B0 h/w.

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I2c83e61fc3b35be01e4581ae309520d89dea260c