IB/hfi: implement PortErrorAction port bounce
PortErrorAction controls which error events will cause the driver to
bounce the port when that error occurs.  PortErrorAction field is
independent of error interrupt-enable CSRs and is set by SM.  It is
cleared on the Offline -> Polling transition.

ExcessBufferOverrun error detection is done by polling RcvBufOvflCnt in
a timer function since there is no interrupt associated with the error.
All other PortErrorAction-checking and bounces are done in the error
interrupt handler functions.

On a PortErrorAction-induced bounce, link down reason is set to indicate
the bounceable error.

INTERNAL: PR126747 - Implement PortErrorAction

Change-Id: I1d63fcf015ed74d18f2f1e457298d0a1429cb98b
Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Brendan Cunningham <brendan.cunningham@intel.com>