nd_btt: handle signed int overflow in acquire_lane
atomic_t is a signed value.  This line:

	*lane = atomic_inc_return(&(btt->last_lane)) % btt->num_lanes;

will result in a bogus lane number once 2^31 is exceeded, which will
then result in a kernel oops when that number is used as an array index.

This patch makes num_lanes a u32.  This causes the int returned by
atomic_int_return() to get promoted to a u32 before the mod
operation, preventing the overflow.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Reported-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>