Ib/hfi: Bring sdma engines to running when port goes active
The sdma engines should not go to the running state until we can actually
service requests. This is not until port active. Note that if we ever want
use sdma for VL15 sma traffic we would need to reverse this.

INTERNAL: PR 125336 - Process hung when unloading the driver after port down

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I0529bec4a39e95ef7536f04e5787f2a408c5aa4c