Ib/hfi/test: Make building an option for regression_runner script.
This change adds an optional argument to not do the build when running the
regression_runner script. Normally this is fine but when trying to run two
sets of tests at the same time from the same directory is problematic since
one test could be rebuilding while the other is trying to load the module.

INTERNAL: PR 125486 - Improve driver regression testing.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I11378b3b227179129f92dfbd5f378c03d1d9ae0a