nd: enforce 4K alignment for namespaces.
Prevent enabling regions that are unaligned, force DPA allocations to
end on a 4K boundary, and validate that previously written labels
conform to 4K alignments.  This is to preclude any need to check for an
I/O-request crossing a DPA boundary in the nd_blk driver, and to
guarantee that pmem.direct_access() need never consider the case of a 4K
mapping running off the end of the namespace.

This assumes that the NFIT driver never runs on an architecture where
the system page size is less than 4K.

Reported-by: Ross Zwisler <ross.zwisler@intel.com>