nd: don't use nsblk->res for displaying the size of a blk namespace
In preparation for some wider block fixes the sizing and the update of
the namespace resource list (nsblk->res array) will be delayed to
__blk_label_update() time.  By delaying the update of the array the
presence of an initialized resource list can be used as a flag that the
labels previously allocated to the namespace need to be torn down.
Otherwise, __blk_label_update() can't determine the old label set from
the latest resource list in the dimm.