Ib/hfi: Ensure PIO send requests get completed by the driver.
This patch fixes two issues where the driver does not generate completion
events for send requests.

The first is when the PIO buffers are full and the
link is donw or otherwise unable to trasmit. It does this by queueing all send
requests after the buffers are full just like before, but now when it goes to
service the request if the link is not active the request will be completed
with an error. This will allow the ULP to get its notification and do error
recovery.

The other issue is due to another bug in the driver where an invalid pkey is
generated. The driver will fail to send the packet after finding the invalid
peky but will never return a completion to the ULP. This patch instead of just
returning a value that will never be looked at, will generate a completion
with an error and allow the ULP to do its own recovery.

INTERNAL: PR 123314 - IB/STL MAD stack will hang when sends do not get a
INTERNAL: completion.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I221f721feea6cd00399cd2826e5804d43006c204