IB/hfi: Remove head lock in packet submission
PSM currently holds a lock around the sdma_send_txreq()
call.  As part of the submission process, the head_lock
might be procurred.

The SDMA code around the PSM complete callback holds the
head_lock as part of the progress can gets the same lock, causing
a deadlock.

This fix eliminates the head_lock in the submission processing
by converting the head_lock to a seqlock and using that
paradyne as to detect progress in a new routine sdma_check_progress
and the "sleep" routine.

Additionally, the flushlist is now protected by a different dedicated
head lock and the flush logic moves flushed txreqs to the a local list
head under the flush lock and calls the complete routine only holding
the head_lock and not holding the flushlist lock.

Other cleanups:
- tx->wait is correctly set in all cases
- The busy callback is removed since no one uses it
- The iowait * is now required and tests are eliminated

INTERNAL: PR 127047 - user sdma causes kernel crashes

Change-Id: I0092b40e61eb160ae4af491e064c0ba116b763de
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>