IB/hfi: fix ahg allocator for verbs/PSM ahg use
The WFR driver includes a check in sdma_ahg_alloc which prevents the
allocation of AHG slots for user contexts if the feature has been
turned off for kernel contexts.

The hfi_cap_mask bitmask can enable/disable AHG for user and kernel
contexts independently. This means that user contexts might use AHG
while kernel context don't.

Since both type of contexts use the same API for allocating AHG slots,
the check that is currently in the ahg allocator has been removed in
favor of independent control in verbs and PSM.

INTERNAL: PR 127125 - WFR driver prevents the use of AHG for user
INTERNAL: context when feature is turned of for kernel contexts

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>