nd_blk: move BW selection to nd_blk_do_bvec
Move the code that does the block window selection from
nd_blk_make_request() down to nd_blk_do_bvec().  This choice requires
knowledge of the DIMM.  The plan is for nd_blk_make_request() to stay in
nd/blk.c, but nd_blk_do_bvec() will move to nd/dimm.c, so only the
latter will have any DIMM knowledge.

This also means that we will only hold a lane for the size of a bio_vec
(one page) instead of for a while bio, which should be better for
latency in situations where there is lane contention.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>