nd: mark label fields that are not applicable to BLK namespaces
'isetcookie': is obvious since there is no related interleave-set for a
block namespace.

'nlabel': while a nice confirmation it becomes to expensive to manage
when updating a BLK namespace.  I.e. it forces an namespace update to
touch all pre-existing labels in the namespace.

'position': similar problem to nlabel, and we always read allocations in
dpa order, so the position of the label is pre-determined by 'dpa'.