nd: export free label space accounting details
Attributes for per-dimm free label space and per-namespace label
allocations.  This helps userspace know if namespace creation will fail,
and the fragmentation level of a blk namespace.  The kernel hides one
available label slot from userspace that it reserves for allowing
updates to a single label.

This also fixes nd_label_nfree() to return the number of possible labels
in the case where the label space is not initialized.