nd: rework BLK label management
There was some confusion in the implementation around when to use nsblk->res
(list of namespaces assigned to a resource) vs nd_mapping->labels (pointers to
labels for all the namespaces on a dimm) vs ndd->dpa (DPA resource tree).

Here are the new rules:

1/ nsblk->res: Contains the 'last' state of the resources for a
namespace.  In the case where a label update fails this will include
pointers to stale resources.  For this reason it can't be used to
determine the current DPA allocation.  That responsibility moves to
scanning ndd->dpa.  The stale resources are useful for determining
"is_old_resource()"

2/ ndd->dpa: The DPA resource tree for the DIMM.  This is always the
most up-to-date location for determining DPA allocations and gets ahead
of the labels until they can be brought up to sync.  If a label update
fails the DPA allocations remain the same.

3/ nd_mapping->labels: Permits converting DPA (start, size) tuples into
label slots so that they can be reaped or maintained over a size update.