nd: don't fail region enabling if at least one namespace successfully registers
nd_region_register_namespaces() may fail to create some child namespace
instances.  In that case the region will succeed in enabling and
userspace can determine that namespaces are missing by 1/ noticing that
a desired block-device with a given UUID is missing 2/ checking
regionX/namespaces in sysfs and observing "<registered
namespaces>/<total available>".  Note, that the <registered namespaces>
in regionX/namespaces is the number of attempted registrations.  The
successfully registered count is determined by counting the number of
regionX/namespaceX.Y instances.