IB/hfi: Reduce the poll restart time
LNI is complete by the time the 8051 firmware indicates
it is ready to accept commands.  There is no need to
wait further.  Reduce the driver restart time from 10s
to 1s as a stepping stone.

Change the comments to give a clearer indication of what
is going on and remove any mention of a specific restart
time.  Move the simulator workaround for the 8051 firmware
wait to inside of wait_fm_ready() as we raise the wait to
the longest firmware wait.

Gather all link start steps into one routine, start_link,
renamed from restart_link.

INTERNAL: PR 127172 - Reduce wait time to restart polling

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>
Change-Id: Ie7d62b4faa607f1c8e8cb64307135b77646d8584