Initial VNIC driver for STL2
Basic infrastructure to register with the network core. The new
module is named opa2_netdev.ko. The skeleton VNIC driver interacts
with the OPA core to obtain a PID and maps the TX/RX command queues.
It then populates a PT entry with an RX buffer and then attempts to
TX some test data into it during the initialization flow as a kernel
client POC. The driver loads and you can see a standard ethernet
interface via ifconfig. It does not handle any socket buffers for
TX or RX but it is primarily meant to be a POC for a kernel client.

Change-Id: Id924f71066c46069f47e50caa3c0d615f96bd256
Reviewed-by: Sanath Kumar <sanath.s.kumar@intel.com>
Reviewed-by: Brian Welty <brian.welty@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>