IB/hfi: Fix bug with Static Rate Control capability bits
The capability bitmask contains matching bits for both
kernel and user contexts, which allow users to enable
features somewhat independently for the two types of
contexts. One of these bits controls Static Rate Control.

However, Static Rate Control (SRC) can not be controlled
completely independently for user and kernel contexts
because not all four combinations of bits are valid.

The one invalid combination is user SRC on and kernel
SRC off. In this case, the driver will automatically
turn the user SRC bit off and as an extra precaution,
scrub the PBC for user SDMA requests.

This commit introduces a new class of capability bits -
one that requires the kernel bit to be enabled in order
to allow the user bit to also be enabled. The SRC bits
are the only ones in this bitmask for now.

INTERNAL: PR127184 - Inconsistency between settings for
INTERNAL: DisallowPbcStaticRateControl in SendCtxt and
INTERNAL: SendDma

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I2f785c0d05f00c6aab6d627feb16ad967120ddb7