nd: btt: optimize lane and map locks
Convert lane locks to spinlocks indexed by a hash of the current cpu
number. Our previous strategy of using a last_lane tracker was causing a
lot of cache-thrash in bouncing both last_lane, and all the spinlocks
among different CPUs caches.

This also adds a check for the number of CPUs, and if we have more lanes
than CPUs, it skips the lane lock entirely in favor of a simple get_cpu
(which implies a preempt_disable).

Another optimization is to pad both lane locks and map locks so that
we only have one lock per cache line. This prevents any false sharing
of cache lines.

Overall, we see a ~30% performance improvement from the above
optimizations.

Signed-off-by: Vishal Verma <vishal.l.verma@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>