IB/hfi: add makesrpm.sh to allow simplication of build processes
Currently there is a separate import script for each WFR git repro
Each repro has its own approach to create a source rpm.  These are documented
on the WFR wiki.  This causes a maintainability issue as it requires 13
different import scripts each with slightly different algorithms to
taking a cloned git repro a producing a srpm.  This problem will only get
worse as it can be anticipate that future revisions may change the process
and hence require release specific versions of those scripts.

The proposed solution is to have a makesrpm.sh script in each repro which
implements the algorithm for building the srpm from a fresh git clone/checkout.
This results in a consistent "API" for use by the import and build to take
any WFR git repro and create a src.rpm.
As a start the change for the WFR driver repro is provided in this patch and
once agreed to, additional makesrpm.sh scripts are ready to be pushed for
9 of the 13 repros.  (last 4 still in development).  As part of this work
the sandbox build process for creating IFS can also be more automated and
simplified to take advantage of these scripts in the existing dobuildall script.

In addition to the makesrpm.sh addition, there is a minor makefile change:
- remove leading ./ in pathnames inside tarball (this confuses the existing
    expand_source script and required the import script to untar and then
    retar withou the ./).  There is no advantage to having ./ in the path.

I chose to include makesrpm.sh in the src.rpm since it will do no harm and
then the src.rpm will represent the complete repo.  However if desired it
could easily be added to the EXCLUDES list

INTERNAL: PR 127255 - improve maintainability of git import process

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Reviewed-by: John Fleck <john.fleck@intel.com>
Signed-off-by: Todd Rimmer <todd.rimmer@intel.com>
Change-Id: I4e52c2e9cafd2e84f0791c0f7c7fee62cca5648f