IB/hfi: Increase total Eager buffer limit to 64MB
The driver had a bug which limitted the total Eager
buffer memory allowed per context to 2MB. This limit
is pretty restrictive so it was necessary to increase
it.

A per-context limit of 64MB is the new limit. The valid
Eager buffer sizes are still limited to powers of two
(2MB, 4MB, 8MB, etc.)

INTERNAL: PR127295 - eager buffer size parameter does
INTERNAL: not work

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I354b4deaa6d6ffd73f5b4e0d2c63dcdc02f4ed80