IB/hfi/test: Check for socket in use before launching test.
Regression tests sometimes fail because a socket can not be set to listneing.
This could happen for instance if the socket is already in use. This patch
addresses the issue by checking for the socket being in use after generating
the random port number.

INTERNAL: PR 127347 - WFR Regression test needs to be tolerant of sockets not
INTERNAL: listening.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>