IB/hfi: fix endian bug in Get(LedInfo)
We're byteswapping the result of a read_csr() prior to interperting
the register content, which is just a dumb bug. Fix that. Also update
the definition of 'struct stl_led_info' to match a recent change to
the spec that requires an additional 32bit reserved field (this is
done to preserve 64bit alignment of the structure when used in
MAD aggregates, for example).

INTERNAL: PR 127238 - Set(LedInfo) returns "Invalid Attribute/Method"

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>