IB/hfi: Account for tid map wrap
Usage of Expected RcvArray entries is tracked
by groups in a set of bitmaps, where each bit
represents a group. The data types of these
bitmaps is "unsigned long" allowing for 64
groups per bitmap.

In the case when the number of Expected RcvArray
groups did not fit nicely into the bits in the
bitmaps (the number of groups is not a multiple
of 64), the driver needs to be able to account
for the left over bits in the last bitmap.

Previously, the driver just did a simple subtraction
from the total number and number of used groups
to get an accurate number of free groups. However,
this was flawed as it allowed for the "unused"
groups to float around the bitmaps resulting in
a completely wrong view of the state of the bitmaps.

With this change, the "unused" groups get marked
as permanently used when the context is being
allocated. This prevents the expected setup and free
functions from even using these bits. In turn, this
provides a much better view of the state of the
allocated/free groups.

Another improvement is come in the change that allows
the driver to cycle through all the bitmaps and check
the starting bitmap one more time at the end. This
ensures that the driver will use any free groups that
are at the beginning of the bitmap where the driver
started.

To help track the tidmap usage, a new trace event is
introduced which shows the state of the maps at setup
and free of expected RcvArray groups.

INTERNAL: PR127124 - mpi_stress with 160 ranks/node
INTERNAL: hangs after message size of 128K

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I7f89be413c3e8c9700ecde3f0915493cb3157099