IB/hfi: Fix a driver memory leak in User SDMA code
The driver was leaking a page of memory for every
SDMA request, which eventually was causing a kernel
crash due to the huge amount of memory leaked.

Also, unpin_vector_pages() can not be inlined as
it is passed as a function pointer.

INTERNAL: PR127165 - Kernel crash with long run of
INTERNAL: mpi_stress with 16 SDMA engines

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Ie78ae3ebaa779a21f0d9c6a0df69b0cba47d2828