IB/hfi: Improve shared context matching
The driver matches shared contexts to the
master context by comparing some values
supplied by the user process against the
same values for the already allocated
contexts.

However, the values that the driver was
comparing were not sufficient to accurately
do the matching. The problem case was
subctxt_id of '0'.

Since the default value for the variable
in the driver's context structures is '0',
the driver was incorrectly matching a non-
shared context to a shared one.

This fix improves the matching by also
including the subctxt_cnt variable in the
matching criteria.

INTERNAL: PR127396 - MPI w/ PSM app fails
INTERNAL: with -n above 2x user send context
INTERNAL: count

Reviewed-by:  Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Id693937fc7e5950d91f87b5e4a72f139017a6586