IB/hfi: Fix part of PBC that gets sanitized
Commit c8706d866b1ee2cbcfbf138d5f249f448b4904c9
added code that was supposed to "sanitize" the
PBC Static Rate Control field provided by user
level if the STATIC_RATE_CTRL feature bit was
not enabled for user contexts.

Unfortunately, the element that was getting
"sanitized" was incorrect and the code ended up
clearing other fields, which should not have
been touched.

This fix corrects that error by changing the
element which is cleaned up to the 3 16bit
value in the PBC.

INTERNAL: PR127184 - Inconsistency between
INTERNAL: settings for DisallowPbcStaticRateControl
INTERNAL: in SendCtxt and SendDma

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Ibf5ba61764e10185de11677aa141628f948cc956